Add aws ec2 inventory
Allow playbook servers to be read from aws ec2.

Inventory can be configured using:

```
inventory "aws", "ec2" {
  connect_on = "PublicIpAddress"
  access_key_id = "access_key_id"
  secret_key = "secret_key"
  region = "us-east-1"
}
```

`connect_on` specifies the address that will be used to connect to the
server. Acceptable values are `PublicIpAddress`, `PublicDnsName`,
`PrivateIpAddress` and `PrivateDnsName`. This default to
`PublicIpAddress`

Currently it will only search 1 region at a time.

`access_key_id` and `secret_key` are not required and will be read from
environment variables or the users `.aws/credentials` file.

This PR also removes the plugin loading code for current simplicity.
This will be restored in later itererations.