Allow roles to have dependents roles
Dependents roles can be defined in a playbook using the dependents
attribute e.g.

```
role "elastic_search" {
  ...
  dependents = ["install_java"]
}
```

This will treverse the list of dependents and create a list of tasks to
run.

Roles will be duplicated if they are used in multiple roles.

The run will fail if there is a dependency loop and stop the run
immediately