allow including files in main role
Introduce `include` block for roles specified in folders. It also set
the convesion that roles within the roles folder must start with
main.yml or no role will be loaded.

Includes are handled relative to the main.yml file of the role and will
only apply to folders within the role folder.