Return data from task and use value in roles
Tasks can now return a map of values as a result that can be used as
variables for subsiquante tasks.

Variables can be accessed within a task as `${
name-of-task.return-value-name}`

In order to allow for other tasks to use the variables from other tasks
the evaluation of tasks have been moved to just before they are run.
Also the returned variables are tied to the  runner for those tasks so
that different servers will have different list of results based on
there own environment