Allow tasks to be marked as conditional
This is a first pass at conditionals for tasks. Conditionals can be
defined by applying the `condition` attribute to tasks.

They are currently limited to evaluating a `true` or `false` string value
returned by a task or set in a variable. They must be wrapped in a `not`
or `is` function within the expression.

For example

```
task "dpkg.package-details" "ruby-package" {
  package = "ruby"
}

task "apt-get" "install-ruby" {
  package = "ruby"
  condition = "${not(ruby-package.installed)}"
}
```