Scope variables to host and roles
Host and roles now have there own variable set an hosts are now unpacked
as a child of role.

Roles now contain there own variables

Parse variables based on roles. Roles can access variables from the host
but not other roles.

In order to keep scope variables for each role create a child scope for
each role. A child scope contains the merge of the parent global
variables with the variables for the role.

Initial runner variables are are a merge of playbook variables and role
variables